[#5ec6ej] If a default PricingTemplate does not exist for an FBO, prompt them to choose one upon login (if a PricingTemplate record exists for their account).

+++ FBOLinx.Web/FBOLinx.Web/Controllers/CustomCustomerTypesController.cs
@@ -159,7 +159,11 @@ namespace FBOLinx.Web.Controllers
                     newDefault.Default = true;
                 }
 
-                _context.PricingTemplate.Remove(currentDefault);
+                if(currentDefault != null)
+                {
+                    _context.PricingTemplate.Remove(currentDefault);
+                }
+                
                 _context.PricingTemplate.Update(newDefault);
 
                 _context.SaveChanges();
